requirements.txt
# Core Framework
streamlit>=1.32.0

# API Clients
requests>=2.31.0
google-genai>=1.0.0

# Environment Management
python-dotenv>=1.0.0
